Moved stock management module screen code to new file
